Support Atom's thr:in-reply-to
RFC 4685 describes the thr:in-reply-to element, which is similar to
In-Reply-To and References for email. This changes the generation
procedure for message-ids and adds a References header to support this.

The use case is mainly blog comments, which are often offered as
separate feeds, and when converted to email should logically be replies
to the corresponding resource.