Implement CGI configuration page
This implements a CGI configuration page, which can be run from the
supplied wrapper script r2e.cgi. It requires running a webserver, so
it's only for people who are already running rss2email on a server. It's
implemented as an extra "cgi" command.

There are two main limitations:
- It requires sudo to be installed, so it's really only for Unix systems
at the moment.

- It only does adding and removing feeds